Rename cmd/bride/server package to cmd/bridge/httpserver
