+++ cmd/bridge/httpserver/redirect-server.go
@@ -1,4 +1,4 @@
-package server
+package httpserver
 
 import (
 	"flag"
